pyproject.toml
[project]
name = "crewai-reels-agent"
version = "0.1.0"
description = "CrewAI workflow for turning long-form videos into reel-style vertical clips."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "crewai>=0.114.0",
  "pyyaml>=6.0.2",
  "openai>=1.68.0",
  "google-api-python-client>=2.164.0",
  "google-auth>=2.38.0",
  "google-auth-oauthlib>=1.2.1",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
